add some changes

+++ src/main/java/com/lead/dashboard/domain/Role.java
@@ -11,7 +11,7 @@ import jakarta.persistence.Table;
 public class Role {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer id;
+  private Long id;
 
 
   private String name;
@@ -23,12 +23,15 @@ public class Role {
   public Role(String name) {
     this.name = name;
   }
+  
+  
 
-public Integer getId() {
+
+public Long getId() {
 	return id;
 }
 
-public void setId(Integer id) {
+public void setId(Long id) {
 	this.id = id;
 }
